madee some changes
